refactor: split onboard guided render
